atm-api: Require both ids for delete/update in customer_has_account

+++ atm-api/models/crud/customer_has_account.js
@@ -25,34 +25,13 @@ module.exports = {
             [ customer_has_account.customerId, customer_has_account.accountId ]
         );
     },
-    delete: (key, id) => {
-        if(typeof(key) === "number") {
-            return db.query(
-                "DELETE FROM customer_has_account WHERE customerId=? AND accountId=?",
-                [ key, id ]
-            );
-        }
-
-        return db.query(
-            key === "customer" ?
-                "DELETE FROM customer_has_account WHERE customerId=?" :
-                "DELETE FROM customer_has_account WHERE accountId=?",
-            [ id ]
-        );
+    delete: (customer, account) => {
+        return db.query("DELETE FROM customer_has_account WHERE customerId=? AND accountId=?", [ customer, account ]);
     },
-    update: (key, id, customer_has_account) => {
-        if(typeof(key) === "number") {
-            return db.query(
-                "UPDATE customer_has_account SET customerId=?, accountId=? WHERE customerId=? AND accountId=?",
-                [ customer_has_account.customerId, customer_has_account.accountId, key, id ]
-            );
-        }
-
+    update: (customer, account, customer_has_account) => {
         return db.query(
-            key === "customer" ?
-                "UPDATE customer_has_account SET customerId=?, accountId=? WHERE customerId=?" :
-                "UPDATE customer_has_account SET customerId=?, accountId=? WHERE accountId=?",
-            [ customer_has_account.customerId, customer_has_account.accountId, id ]
+            "UPDATE customer_has_account SET customerId=?, accountId=? WHERE customerId=? AND accountId=?",
+            [ customer_has_account.customerId, customer_has_account.accountId, customer, account ]
         );
     }
 };
